Fix problem with spaces in variables
systemd quotes environment variables so they're passed as a single
parameter even if they contain spaces. As the moderation bots can't
handle that for the provided room names, this adds a workaround by not
passing the rooms as space separated list via an environment variable,
but as a space separated list in the exec command instead.