Use correct shaper for bots
Some bots, especially XpartaMuPP send a lot of stanzas. Therefore, they
need a to have a much higher rate of messages as the default shaper
allows. While this was already correctly configured on the old lobby VM,
configuring a different shaper wasn't honored as part of the migeation.
This commit re-adds it.