Use ansible_local provisioner with Vagrant
This switches the provisioner used when creating a local lobby VM with
Vagrant from "ansible" to "ansible_local". This makes it easier for
users on Windows to deploy a local lobby VM, as they don't need to
install Ansible anymore, which isn't available for Windows natively.

As Ansible is now running inside the VM, the VM does require more memory
than before, however with 1GB that should still be low enough to run on
pretty much any computer in use nowadays.