Increase traffic rate for bots
As it turns out 200KB/s isn't enough bandwidth for the bots to send out
updates when many players are online. Instead they appear to be laggy,
as they send out updates for game lists and ratings delayed. To avoid
this laggyness while still keeping some shaping for the bots in place,
this increases the available for bots to 1MB/s, with a burst rate of
10MB/s. While this fixes this problem for now, moving to PubSub
would be a more elegant solution.