Ensure ejabberd is running before running commands
The Ansible playbook contains several tasks which require ejabberd to be
running to succeed. However, the playbook didn't ensure that ejabberd
was running, so manually stopping ejabberd prior to running the playbook
would result in errors. This commit fixes that by ensuring ejabberd is
enabled and started before running the tasks which require it to run.