Allow unlimited bandwidth for bots again
This restores the behavior prior to the lobby migration of granting the
bots unlimited bandwidth for sending XMPP stanzas. That's necessary, as
XpartaMuPP has to send a lot of data when many players are online and
many games get hosted. Prior tries increasing the existing rate limit
didn't succeed as XpartaMuPP needs that much traffic in such a situation
that traffic shaping in general doesn't make much sense.

While this change fixes this problem for now, moving to PubSub would be
a more elegant solution.