Configure unattended-upgrades
This adds unattended upgrades of apt packages to keep them secure and
up-to-date without manual intervention.

To avoid disruption of the 0 A.D, multiplayer lobby, ejabberd is
excluded from unattended upgrades. As unattached-upgrades only handles
apt packages, the lobby bots and the Python third-party libraries they
use are also not affected by this.