Change the naming of lobby config files
This changes the names of the configuration files for the lobby to
include the name of the host they're for in their filename. This does
open the possibility to deploy multiple multiplayer lobbies from the
same code checkout and makes it more explicit which configuration is
being used during deployment.

Users with a customized configuration will have to rename their local
configuration files as a result of these changes to match the new
naming scheme. As these changes also change the logical name used for
the Vagrant VM from `default` to `lobby-vagrant`, already deployed
Vagrant VM's will need to be recreated.