Use online backup API for SQLite backups
This improves how backups for SQLite databases are done. Instead of
dumping the database content into a text file, using the "dump" command,
a copy of the actual database is now being done by by the "backup"
command. This is much faster than using the "dump" command.
In addition this sets a busy timeout of 10s before doing the backup to
avoid failures when the database is briefly locked by another process
when the backup starts.