Limit the syscalls available to the lobby bots
This reduces the attack surface of the bots, as it limits the
system calls they're allowed to use to the ones they actually
need. Attackers gaining remote code execution will be limited to
these system calls as well and can't exploit potential security
vulnerabilities in other system calls to elevate privileges.