Set a more restrictive default umask
By default Debian uses a umask of 022, which results in newly created
files being world readable. That's not desired for our use case, as we
want to restrict access of different users, especially the lobby bots,
as much as possible. Therefore, this changes the default umask from 022
to 077 to ensure only the creator of a file can access it by default.

When updating an existing lobby server, the file permissions for already
existing lobby bots have to be updated manually to make use of the more
restrictive permissions.