Use config files for the lobby bots
Instead of using environment variables for configuration, this switches
to using configuration files for the configuration of the lobby bots.
This ensures that that sensitive configuration values aren't visible in
the process list anymore.

Deploying this requires adjusting the lobby configuration files to
account for the changed key names. An example for that is included in
the changes to `config-lobby-vagrant.yml` as part of this commit.